Set up dummy page to test navigation and created base adapter
