+++ app/src/main/res/layout/fragment_locations_map.xml
@@ -4,7 +4,8 @@
     xmlns:tools="http://schemas.android.com/tools"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    tools:context=".ui.location.LocationsMapFragment">
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/googleMap"
