+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -32,6 +32,14 @@
         android:id="@+id/nav_locations"
         android:name="com.skysphere.skysphere.ui.location.LocationsFragment"
         android:label="@string/menu_locations"
+        tools:layout="@layout/fragment_locations" />
+
+
+
+    <fragment
+        android:id="@+id/nav_map"
+        android:name="com.skysphere.skysphere.ui.location.LocationsMapFragment"
+        android:label="@string/menu_locations"
         tools:layout="@layout/fragment_locations_map" />
 
     <fragment
